% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFit.R
\name{extractFit}
\alias{extractFit}
\title{Extract topic quality statistics from a `manyTopics` fitted object}
\usage{
extractFit(models, documents, M = 10, n = 20)
}
\arguments{
\item{models}{A list returned by `stm::manyTopics`}

\item{documents}{The documents to be modeled. Object must be a list of with each element corresponding to a document. Each document is represented as an integer matrix w\ ith two rows, and columns equal to the number of unique vocabulary words in the document. The first row contains the 1-indexed vocabulary entry and the sec\ ond row contains the number of times that term appears. Corpora can be imported using the reader function and manipulated using the prepDocuments.}

\item{M}{The number of word tokens to calculate semantic coherence and exclusivity statistics from.}

\item{n}{The number of exemplars returned by `stmQuality::findThoughts0` to estimate thetas from.}
}
\value{
A data-frame with K, topics, semcoh, exclusivity, theta
}
\description{
Allows customisable semantic coherence, exclusivity, and theta parameters and returns 
an easy to work with data frame
}
\examples{
 fit <- manyTopics(out$documents, out$vocab, K = K,
 prevalence =~ cat * s(age), max.em.its = 100,
 data = out$meta, runs = 3, init.type = "Spectral", verbose = F)
 k <- extractFit(fit, M = 20, n = 5)
}
