% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlot.stm.R
\name{pcaPlot.stm}
\alias{pcaPlot.stm}
\title{Visualise models returned by `stm::manyTopics`}
\usage{
pcaPlot.stm(x)
}
\arguments{
\item{x}{A data-frame returned by `extractFit`}
}
\description{
Visualise models returned by `stm::manyTopics`
}
